% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedFunctions.R
\name{as.data.frame.SDEFSR_Dataset}
\alias{as.data.frame.SDEFSR_Dataset}
\title{S3 function to convert into a data.frame the SDEFSR dataset
 
 This function converts a SDEFSR_Dataset object into a data.frame}
\usage{
\method{as.data.frame}{SDEFSR_Dataset}(x, ...)
}
\arguments{
\item{x}{The \code{SDEFSR_Dataset} object to view}

\item{...}{Additional arguments passed to the as.data.frame function}
}
\value{
a data.frame with the dataset uncoded. Numeric attributes are "numeric" class, while categorical attributes are "factor"
 
 @examples
 
 as.data.frame(habermanTra)
}
\description{
S3 function to convert into a data.frame the SDEFSR dataset
 
 This function converts a SDEFSR_Dataset object into a data.frame
}
\details{
Internally, a \code{SDEFSR_Dataset} object has a list of of examples
 and this examples are coded numerically. This function decode these examples and convert the list into a data.frame
}
