% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{carTra}
\alias{carTra}
\title{Car evaluation dataset}
\format{
A SDEFSR_Dataset class with 1382 instances, 6 variables (without the target variable) and 4 values for the target Variable.
    Three labels for each variable are defined.
}
\source{
M. Bohanec and V. Rajkovic: Knowledge acquisition and explanation for multi-attribute 
    decision making. In 8th Intl Workshop on Expert Systems and their Applications, Avignon, 
    France. pages 59-78, 1988.

B. Zupan, M. Bohanec, I. Bratko, J. Demsar: Machine learning by function decomposition. 
    ICML-97, Nashville, TN. 1997 (to appear).
}
\description{
Training data for the car dataset
}
\details{
Car Evaluation Database was derived from a simple hierarchical decision model. 
     The model evaluates cars according to six input attributes: buying, maint, doors, 
     persons, lug_boot, safety.
}
\examples{
   carTra$data
   carTra$attributeNames
   
}
