% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{changeTargetVariable}
\alias{changeTargetVariable}
\title{Change the targetVariable of a KEEL Dataset}
\usage{
changeTargetVariable(dataset, posVariable)
}
\arguments{
\item{dataset}{The KEEL dataset class}

\item{posVariable}{The position of the variable to set as target Variable.}
}
\value{
The dataset with the variables changed
}
\description{
Change the actual target variable for another one if it is categorical.
}
\examples{
changeTargetVariable(carTra, 3)
\dontrun{
Throws an error because the variable selected is numerical:

changeTargetVariable(habermanTra, 1)
}
}

