% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{createHumanReadableRule}
\alias{createHumanReadableRule}
\title{Creates a human-readable representation of a rule for insert into a 'SDR_Rules' object}
\usage{
createHumanReadableRule(rule, dataset, DNF = FALSE)
}
\arguments{
\item{rule}{The rule we want to get the representation}

\item{DNF}{Logical value indicating if the rule is in DNF format}

\item{FuGePSD}{Logical value indicatin if rule is in the format of the FuGePS algorithm}
}
\value{
A string with the representation of the rule.
}
\description{
This function creates an string representation of a given rule
}

