% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{FUGEPSD}
\alias{FUGEPSD}
\title{Make a subgroup discovery task using the FuGePSD algorithm.}
\usage{
FUGEPSD(paramFile)
}
\description{
Make a subgroup discovery task using the FuGePSD algorithm.
}

