% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{FuGePSD_crossover}
\alias{FuGePSD_crossover}
\title{Performs the FuGePSD crossover function.}
\usage{
FuGePSD_crossover(rule1, rule2, nvars)
}
\arguments{
\item{rule1}{a rule object}

\item{rule2}{another rule object.}

\item{nvars}{number of variables in the dataset INCLUDING class variable.}
}
\value{
a new rule object.
}
\description{
Performs the FuGePSD crossover function.
}

