% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{FuGePSD_Mutation}
\alias{FuGePSD_Mutation}
\title{Perfom the mutation operator for FuGePSD algorithm.}
\usage{
FuGePSD_Mutation(rule, dataset)
}
\arguments{
\item{rule}{The rule object to be mutated.}

\item{dataset}{a keel object associated with the rule.}
}
\value{
a new rule object with the rule mutated.
}
\description{
Perfom the mutation operator for FuGePSD algorithm.
}

