% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{habermanRules}
\alias{habermanRules}
\title{Haberman survival rule set}
\source{
Haberman, S. J. (1976). Generalized Residuals for Log-Linear Models, 
    Proceedings of the 9th International Biometrics Conference, Boston, pp. 104-122.

Landwehr, J. M., Pregibon, D., and Shoemaker, A. C. (1984), Graphical Models for 
    Assessing Logistic Regression Models (with discussion), Journal of the American Statistical 
    Association 79: 61-83.

Lo, W.-D. (1993). Logistic Regression Trees, PhD thesis, Department of Statistics,
    University of Wisconsin, Madison, WI.
}
\description{
Rules generated by the SDIGA algorithm with the default parameters for the \code{haberman} dataset.
}
\details{
The rule set contains only two rules. One for each target variable
}
\examples{
    habermanRules
}
