% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{habermanTst}
\alias{habermanTst}
\title{Haberman survival data set}
\format{
A SDEFSR_Dataset class with 62 instances, 3 variables (without the target variable) and 2 values for the target variable.
    Three fuzzy labels for each numerical variable are defined.
}
\source{
Haberman, S. J. (1976). Generalized Residuals for Log-Linear Models, 
    Proceedings of the 9th International Biometrics Conference, Boston, pp. 104-122.

Landwehr, J. M., Pregibon, D., and Shoemaker, A. C. (1984), Graphical Models for 
    Assessing Logistic Regression Models (with discussion), Journal of the American Statistical 
    Association 79: 61-83.

Lo, W.-D. (1993). Logistic Regression Trees, PhD thesis, Department of Statistics,
    University of Wisconsin, Madison, WI.
}
\description{
Test data for the Haberman dataset.
}
\details{
This data set contains cases from a study that was conducted between 1958 and 1970 
    at the University of Chicago's Billings Hospital on the survival of patients who had undergone 
    surgery for breast cancer. The task is to determine if the patient survived 5 years or longer 
    (positive) or if the patient died within 5 year (negative)
}
\examples{
    habermanTra$data 
    habermanTra$fuzzySets
}
