% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{keelFromARFF}
\alias{keelFromARFF}
\title{Reads an ARFF file}
\usage{
keelFromARFF(file)
}
\arguments{
\item{file}{The ARFF file to read.}
}
\value{
a 'keel' object ready to use with the algorithms that are in the package
}
\description{
This function reads an ARFF file and get the subyacent \code{keel} object
}

