% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{modifyFuzzyCrispIntervals}
\alias{modifyFuzzyCrispIntervals}
\title{Modifiy the number of Fuzzy Labels of the dataset.}
\usage{
modifyFuzzyCrispIntervals(dataset, nLabels)
}
\arguments{
\item{dataset}{The dataset to modify their fuzzy labels definitions}

\item{nLabels}{The new number of fuzzy labels.}
}
\value{
This function returns the same dataset with their fuzzy definitions modified.
}
\description{
This function change the number of fuzzy labels defined in the current dataset.
}

