% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{orderRules}
\alias{orderRules}
\title{Return an ordered rule set by a given quality measure
 
 This function orders a rule set in descendant order by a given quality measure that are available on the object}
\usage{
orderRules(ruleSet, by = "FuzzyConfidence")
}
\arguments{
\item{ruleSet}{The rule set passed as a \code{SDR_Rules} object}

\item{by}{a String with the name of the quality measure to order by. Valid values are: \code{Coverage, Unusualness, Significance, FuzzySupport, FuzzyConfidence, CrispConfidence, Tpr, Fpr}.}
}
\value{
another \code{SDR_Rules} object with the rules ordered
}
\description{
Return an ordered rule set by a given quality measure
 
 This function orders a rule set in descendant order by a given quality measure that are available on the object
}
\examples{
 orderRules(habermanRules)
 
}

