% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{parseTime}
\alias{parseTime}
\title{Parse a time differente to "x hours, y minutes and z seconds"}
\usage{
parseTime(actual, initial)
}
\arguments{
\item{actual}{End time in UNIX int format (i.e. as.numeric(Sys.time()))}

\item{initial}{Initial time in Unix format.}
}
\value{
A human-readable string with time difference.
}
\description{
Parse a time differente to "x hours, y minutes and z seconds"
}

