% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedFunctions.R
\name{plot.SDEFSR_Rules}
\alias{plot.SDEFSR_Rules}
\title{Plot a rule set generated by a SDEFSR algorithm}
\usage{
\method{plot}{SDEFSR_Rules}(x, ...)
}
\arguments{
\item{x}{an \code{SDEFSR_Rules} object generated by a subgroup discovery algorithm of the SDEFSR package}

\item{...}{additional arguments passed to the plot}
}
\value{
A TPR vs FPR graph generated by ggplot2
}
\description{
This function plots the rule set by means of a bar graph that shows TPR vs FPR quality measure of each rule
}
\details{
This function works depending on the package ggplot2 that allow to generate such graph. If the package ggplot2 is not 
installed, the this function ask the user to install it. After install, load the package and show the graph.

 A TPR vs FPR graph show the precision of a rule. Quality rules has big TPR values and small FPR values.
 Big values of both quality measures indicates that the rule is too much general and it is too obvious. 
 Small values of both indicates that the rule is too much specific and it would be an invalid rule.
}
\examples{
plot(habermanRules)

}
