% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Pop.fuzzyReasoningMethod}
\alias{Pop.fuzzyReasoningMethod}
\title{Calcultes the Fuzzy Reasoning Method for all examples for a given rule population}
\usage{
Pop.fuzzyReasoningMethod(pop, dataset, examplesNoClass, frm, categorical,
  numerical, t_norm)
}
\arguments{
\item{pop}{A list with all rule object that define the population}

\item{dataset}{the keel object with the dataset information.}

\item{examplesNoClass}{matrix with all examples of the dataset without the class attribute. One examples per column.}

\item{frm}{An integer specifing the tipo of fuzzy reasoning method to use. 0 for Winning Rule, 1 for Normalized Sum and 2 for Arithmetic Mean.}

\item{categorical}{Logical vector indicating which attributes of the dataset are categorical.}

\item{numerical}{Logical vector indicating which attributes of the dataset are numerical.}

\item{t_norm}{The t_norm to use. 0 for minimum t-norm. 1 for product t-norm.}
}
\value{
a vector indicating the class predicted for each example.
}
\description{
Calcultes the Fuzzy Reasoning Method for all examples for a given rule population
}

