% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{print.keel}
\alias{print.keel}
\title{S3 function to print in console the contents of the dataset}
\usage{
\method{print}{keel}(x, ...)
}
\arguments{
\item{x}{The \code{keel} object to view}

\item{...}{Additional arguments passed to the print function}
}
\value{
a matrix with the dataset uncoded.
}
\description{
This function shows the matrix of data uncoded.
}
\details{
This function show the matix of data. Internally, a \code{keel} object has a list of of examples
 and this examples are coded numerically. This function decode these examples and convert the list into a matrix.
}

