% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedFunctions.R
\name{print.SDEFSR_Dataset}
\alias{print.SDEFSR_Dataset}
\title{S3 function to print in console the contents of the dataset
 
 This function shows the matrix of data uncoded.}
\usage{
\method{print}{SDEFSR_Dataset}(x, ...)
}
\arguments{
\item{x}{The \code{SDEFSR_Dataset} object to view}

\item{...}{Additional arguments passed to the print function}
}
\value{
a matrix with the dataset uncoded.
 
 @examples
 
 print(habermanTra)
}
\description{
S3 function to print in console the contents of the dataset
 
 This function shows the matrix of data uncoded.
}
\details{
This function show the matix of data. Internally, a \code{SDEFSR_Dataset} object has a list of of examples
 and this examples are coded numerically. This function decode these examples and convert the list into a matrix.
}
