% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{printFuGePSDParameters}
\alias{printFuGePSDParameters}
\title{Prints parameters information aboute the execution of the algorithm in console and in the output file
specified by this parameters.}
\usage{
printFuGePSDParameters(parameters, dataset, inputFromFiles = TRUE)
}
\arguments{
\item{parameters}{A list with all neccesary parameters.}

\item{dataset}{A \code{keel} object with all the information about the dataset.}

\item{inputFromFiles}{A logical indicating if the input datasets are given from a file or from an object in the R environment.}
}
\description{
Prints parameters information aboute the execution of the algorithm in console and in the output file
specified by this parameters.
}

