% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leerDatos.R
\name{read.keel}
\alias{read.keel}
\title{Reads a KEEL format file}
\usage{
read.keel(file, nLabels = 3)
}
\arguments{
\item{file}{The path of the file in KEEL format}

\item{nLabels}{The number of fuzzy labels to create on numerical variables.}
}
\description{
This function reads a KEEL dataset file and store the information
  in a keel class. This function also create triangular fuzzy sets definitions
  for execute in SDIGA, MESDIF or NMEEF-SD algorithms
}
\details{
The KEEL format file has a fixed format, if the file hasn´t got that format, it launch an error
}
\author{
Ángel M. García-Vico <amgv0009@red.ujaen.es>
}
\seealso{
KEEL Software Tool: \url{http://www.keel.es}
}

