% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.addLabel}
\alias{Rule.addLabel}
\title{Adds a new random label to the variable selected in the rule.}
\usage{
Rule.addLabel(rule, variable, dataset)
}
\description{
@param rule The rule to work with
 @param variable the selected variable (is the index of the list $antecedent of the rule)
 @param dataset, the dataset where getting all the data.
 
 @return the same rule but with the params reinitialized and marked as non evaluated.
}

