% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.addVariable}
\alias{Rule.addVariable}
\title{Add a new variable to the antecedent of the rule.}
\usage{
Rule.addVariable(rule, dataset, DNF)
}
\arguments{
\item{rule}{The rule to work with.}

\item{dataset}{The dataset to get the data.}

\item{DNF}{Rule representation. If TRUE, dnf representation, else, canonica.}
}
\value{
A new rule with an added variable.
}
\description{
Add a new variable to the antecedent of the rule.
}

