% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.changeLabel}
\alias{Rule.changeLabel}
\title{Change randomly a label associated to the rule with a non-existing label on the rule}
\usage{
Rule.changeLabel(rule, variable)
}
\description{
@param rule The rule to work with
 @param variable the selected variable (is the index of the list $antecedent of the rule)
}

