% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.clearAntecedent}
\alias{Rule.clearAntecedent}
\title{Remove completely the antecedent part of a rule.
 
 @param rule The rule to where we want to remove the antecedent.
 
 @return A new rule with an empty antecedent}
\usage{
Rule.clearAntecedent(rule)
}
\description{
Remove completely the antecedent part of a rule.
 
 @param rule The rule to where we want to remove the antecedent.
 
 @return A new rule with an empty antecedent
}

