% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.evaluate}
\alias{Rule.evaluate}
\title{Evaluate a single rule. 
 
 @param rule The rule we want to evaluate (Class "Rule").
 @param dataset The keel dataset object with the examples to compare with the rule (Class "keel")
 @param data Matrix with the data of the dataset, one colum per rule. The data must not contain the last column, the class. (use .separate for this task and convert the list into a matrix)
 @param categoricalValues a logical vector indicating which attributes in the dataset are categorical
 @param numericalValues a logical vector indicating which attributes in the dataset are numerical
 @param t_norm The T-norm to use. 0 for minimum t-norm, 1 for product t-norm (default: 1)
 @param ruleWeight An integer with the rule weighting method. \itemize{
        \item 0 -> Classic Certainty Factor weight
        \item 1 -> Penalized Certainty Factor weight II
        \item 2 > Penalized Certainty Factor weight IV
        \item 3 -> No Rule Weight
        }}
\usage{
Rule.evaluate(rule, dataset, data, categoricalValues, numericalValues,
  t_norm = 1, ruleWeight = 0)
}
\value{
The rule evaluated.
}
\description{
Evaluate a single rule. 
 
 @param rule The rule we want to evaluate (Class "Rule").
 @param dataset The keel dataset object with the examples to compare with the rule (Class "keel")
 @param data Matrix with the data of the dataset, one colum per rule. The data must not contain the last column, the class. (use .separate for this task and convert the list into a matrix)
 @param categoricalValues a logical vector indicating which attributes in the dataset are categorical
 @param numericalValues a logical vector indicating which attributes in the dataset are numerical
 @param t_norm The T-norm to use. 0 for minimum t-norm, 1 for product t-norm (default: 1)
 @param ruleWeight An integer with the rule weighting method. \itemize{
        \item 0 -> Classic Certainty Factor weight
        \item 1 -> Penalized Certainty Factor weight II
        \item 2 > Penalized Certainty Factor weight IV
        \item 3 -> No Rule Weight
        }
}

