% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.toCANVectorRepresentation}
\alias{Rule.toCANVectorRepresentation}
\title{Converts a rule antecedent to a CANONICA vector representation}
\usage{
Rule.toCANVectorRepresentation(rule, dataset)
}
\arguments{
\item{rule}{The rule that we want to evaluate.}

\item{dataset}{The keel dataset which rule refers to.}
}
\value{
a matrix with one column to use easily with fitness functions of this package
}
\description{
The function converts a rule into a CANONICA vector representation for ease the evaluation.
The evaluation of a rule with a vetor representation can be evaluated througth functions
of evaluation of rule (.fit13 or .fitnessMESDIF) that are available inside the package SDR
}

