% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{SDR_DatasetFromARFF}
\alias{SDR_DatasetFromARFF}
\title{Reads an ARFF file}
\usage{
SDR_DatasetFromARFF(file)
}
\arguments{
\item{file}{The ARFF file to read.}
}
\value{
a 'SDR_Dataset' object ready to use with the algorithms that are in the package
}
\description{
This function reads an ARFF file and get the subyacent \code{SDR_Dataset} object
}

