% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{SDR_DatasetFromCSV}
\alias{SDR_DatasetFromCSV}
\title{Reads a CSV file and return a SDR_Dataset object to be executed by an algorithm of the SDR package}
\usage{
SDR_DatasetFromCSV(file, relation_name, sep = ",", quote = "\\"",
  dec = ".", na.strings = "?")
}
\arguments{
\item{file}{The path to the csv file.}

\item{relation_name}{The name of the relation to use}

\item{sep}{Separator used to separate between values in the file, as it is csv, the default is ","}

\item{quote}{The character used to identify strings in the csv file}

\item{dec}{The character used to identify decimal values. By default it is "."}

\item{na.strings}{The character used to identify lost data. By default it is "?"}
}
\value{
An \code{SDR_Dataset} object that contains all neccesary information to execute the algorithms
}
\description{
Reads a CSV file and return a SDR_Dataset object to be executed by an algorithm of the SDR package
}
\author{
Angel M. Garcia <agvico@ujaen.es>
}

