% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedFunctions.R
\name{sort.SDEFSR_Rules}
\alias{sort.SDEFSR_Rules}
\title{@title Return an ordered rule set by a given quality measure
 
 @description This function sorts a rule set in descendant order by a given quality measure that are available on the object}
\usage{
\method{sort}{SDEFSR_Rules}(ruleSet, by = "Confidence", decreasing = TRUE)
}
\arguments{
\item{ruleSet}{The rule set passed as a \code{SDEFSR_Rules} object}

\item{by}{a String with the name of the quality measure to order by. Valid values are: \code{nVars, Coverage, Unusualness, Significance, FuzzySupport, Support, FuzzyConfidence, Confidence, Tpr, Fpr}.}

\item{decreasing}{A logical indicating if the sort should be increasing or decreasing. By default, decreasing.}
}
\value{
another \code{SDEFSR_Rules} object with the rules sorted
}
\description{
@title Return an ordered rule set by a given quality measure
 
 @description This function sorts a rule set in descendant order by a given quality measure that are available on the object
}
\examples{
 sort(habermanRules)
 
}

