% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{summary.keel}
\alias{summary.keel}
\title{S3 function to summary a keel object}
\usage{
\method{summary}{keel}(object, ...)
}
\arguments{
\item{object}{A \code{keel} class.}

\item{...}{Additional arguments to the summary function.}
}
\description{
Summary relevant data of a \code{keel} dataset.
}
\details{
This function show important information about the \code{keel} dataset for the user. Note that it does not
show all the information available. The rest is only for the algorithms. The values that appear are accesible by the
\code{$} operator, e.g. dataset$relation or dataset$examplesPerClass.
}
\examples{
summary(carTra)
}

