% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedFunctions.R
\name{summary.SDEFSR_Dataset}
\alias{summary.SDEFSR_Dataset}
\title{S3 function to summary a SDEFSR_Dataset object}
\usage{
\method{summary}{SDEFSR_Dataset}(object, ...)
}
\arguments{
\item{object}{A \code{SDEFSR_Dataset} class.}

\item{...}{Additional arguments to the summary function.}
}
\description{
Summary relevant data of a \code{SDEFSR_Dataset} dataset.
}
\details{
This function show important information about the \code{SDEFSR_Dataset} dataset for the user. Note that it does not 
show all the information available. The rest is only for the algorithms. The values that appear are accessible by the
\code{$} operator, e.g. dataset$relation or dataset$examplesPerClass.
}
\examples{
 
summary(carTra) 

}
