% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{summary.SDR_Dataset}
\alias{summary.SDR_Dataset}
\title{S3 function to summary a SDR_Dataset object}
\usage{
\method{summary}{SDR_Dataset}(object, ...)
}
\arguments{
\item{object}{A \code{SDR_Dataset} class.}

\item{...}{Additional arguments to the summary function.}
}
\description{
Summary relevant data of a \code{SDR_Dataset} dataset.
}
\details{
This function show important information about the \code{SDR_Dataset} dataset for the user. Note that it does not 
show all the information available. The rest is only for the algorithms. The values that appear are accesible by the
\code{$} operator, e.g. dataset$relation or dataset$examplesPerClass.
}
\examples{
 
summary(carTra) 

}

