% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{tokenCompetition}
\alias{tokenCompetition}
\title{Runs a token competition procedure.}
\usage{
tokenCompetition(pop, dataset)
}
\arguments{
\item{pop}{A list of 'Rule' objects.}

\item{dataset}{A 'keel' object with all the information about the dataset we are working.}
}
\value{
A list of 'Rule' objects with that rules that pass this token competition procedure.
}
\description{
Runs a token competition procedure.
}

