% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{tournamentSelection}
\alias{tournamentSelection}
\title{Performs tournament selection for the FuGePSD algorithm}
\usage{
tournamentSelection(pop, tamTournament)
}
\arguments{
\item{pop}{The rule population}

\item{tamToutnament}{The size of the tornament (>= 2)}
}
\description{
It makes a tournament selection for the FuGePSD algorithm with variable tournament size.
}

