% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{writeRuleFile}
\alias{writeRuleFile}
\title{Writes rules into a human-readable format into a file.}
\usage{
writeRuleFile(pop, dataset, fileName)
}
\arguments{
\item{pop}{A list of 'Rule' objects with the population we want to save.}

\item{dataset}{A 'keel' object with all the information about the dataset the rules are pointing.}

\item{fileName}{String with the path to store the rules.}
}
\description{
Writes rules into a human-readable format into a file.
}
\details{
This function overwrites if the file specified by \code{fileName} exists. Be careful!
}

