% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{writeTestQMFile}
\alias{writeTestQMFile}
\title{Writes rules into a human-readable format into a file.}
\usage{
writeTestQMFile(pop, fileName, numExamples)
}
\arguments{
\item{pop}{A list of 'Rule' objects with the population we want to save.}

\item{fileName}{String with the path to store the rules.}

\item{numExamples}{An integer with the number of examples in the test file.}
}
\description{
Writes rules into a human-readable format into a file.
}
\details{
This function overwrites if the file specified by \code{fileName} exists. Be careful!
}

