% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{changeTargetVariable}
\alias{changeTargetVariable}
\title{Change the target Variable of a \code{'keel'} Dataset}
\usage{
changeTargetVariable(dataset, variable)
}
\arguments{
\item{dataset}{The KEEL dataset class}

\item{variable}{The position (or the name) of the variable to set as target Variable.}
}
\value{
The dataset with the variables changed
}
\description{
Change the actual target variable for another one if it is categorical.
}
\examples{
\dontrun{
changeTargetVariable(carTra, 3)
changeTargetVariable(carTra, "Doors")

Throws an error because the variable selected is numerical:
changeTargetVariable(habermanTra, 1)
}
}

