% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{germanTra}
\alias{germanTra}
\title{German Credit data set}
\format{A keel class with 800 instances, 20 variables (without the target variable)
    and 2 values for the target class.}
\source{
\url{http://sci2s.ugr.es/keel/dataset.php?cod=88}
}
\description{
Training data for the german dataset
}
\details{
A numerical version of the Statlog German Credit Data data set.
    Here, the task is to clasify customers as good (1) or bad (2),
    depending on 20 features about them and their bancary accounts.
}
\examples{
germanTra$data
}

