% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Utils.R
\name{improvedTable}
\alias{improvedTable}
\title{Improved table creation for .get_values6}
\usage{
improvedTable(dataset, classNames)
}
\arguments{
\item{dataset}{A matrix with the data}

\item{classNames}{a vector with the names of the attributes.}
}
\value{
a named vector with the number of instances per class.
}
\description{
Improved table creation for .get_values6
}

