% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.addLabel}
\alias{Rule.addLabel}
\title{Adds a new random label to the variable selected in the rule.}
\usage{
Rule.addLabel(rule, variable, dataset)
}
\arguments{
\item{rule}{The rule to work with}

\item{variable}{the selected variable (is the index of the list $antecedent of the rule)}

\item{dataset,}{the dataset where getting all the data.}
}
\value{
the same rule but with the params reinitialized and marked as non evaluated.
}
\description{
Adds a new random label to the variable selected in the rule.
}

