% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.clearAntecedent}
\alias{Rule.clearAntecedent}
\title{Remove completely the antecedent part of a rule.}
\usage{
Rule.clearAntecedent(rule)
}
\arguments{
\item{rule}{The rule to where we want to remove the antecedent.}
}
\value{
A new rule with an empty antecedent
}
\description{
Remove completely the antecedent part of a rule.
}

