% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.deleteVariable}
\alias{Rule.deleteVariable}
\title{Delete a random variable from a rule}
\usage{
Rule.deleteVariable(rule)
}
\arguments{
\item{rule}{The rule to work with.}
}
\value{
A new rule with a random variable in the antecedent removed.
}
\description{
Delete a random variable from a rule
}

