% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuGePSD.R
\name{Rule.exchangeVariables}
\alias{Rule.exchangeVariables}
\title{Create a new rule by mixing the antecedent part of two rules.}
\usage{
Rule.exchangeVariables(rule1, rule2)
}
\arguments{
\item{rule1}{A rule}

\item{rule2}{Another rule}
}
\value{
a list with the antecedents mixed. (This list would be the $antecedent part of another rule)
}
\description{
Create a new rule by mixing the antecedent part of two rules.
}

